# ColoredConsoleTarget performance improvements.