--- a/src/NLog/Targets/ConsoleWordHighlightingRule.cs
+++ b/src/NLog/Targets/ConsoleWordHighlightingRule.cs
@@ -97,9 +97,23 @@ namespace NLog.Targets
         public bool IgnoreCase { get; set; }
 
         /// <summary>
+        /// Gets or sets the foreground color.
+        /// </summary>
+        /// <docgen category='Formatting Options' order='10' />
+        [DefaultValue("NoChange")]
+        public ConsoleOutputColor ForegroundColor { get; set; }
+
+        /// <summary>
+        /// Gets or sets the background color.
+        /// </summary>
+        /// <docgen category='Formatting Options' order='10' />
+        [DefaultValue("NoChange")]
+        public ConsoleOutputColor BackgroundColor { get; set; }
+
+        /// <summary>
         /// Gets the compiled regular expression that matches either Text or Regex property.
         /// </summary>
-        public Regex CompiledRegex
+        private Regex CompiledRegex
         {
             get
             {
@@ -116,7 +130,7 @@ namespace NLog.Targets
                         }
                     }
 
-                    RegexOptions regexOptions = RegexOptions.Compiled;
+                    RegexOptions regexOptions = RegexOptions.None;
                     if (this.IgnoreCase)
                     {
                         regexOptions |= RegexOptions.IgnoreCase;
@@ -129,21 +143,7 @@ namespace NLog.Targets
             }
         }
 
-        /// <summary>
-        /// Gets or sets the foreground color.
-        /// </summary>
-        /// <docgen category='Formatting Options' order='10' />
-        [DefaultValue("NoChange")]
-        public ConsoleOutputColor ForegroundColor { get; set; }
-
-        /// <summary>
-        /// Gets or sets the background color.
-        /// </summary>
-        /// <docgen category='Formatting Options' order='10' />
-        [DefaultValue("NoChange")]
-        public ConsoleOutputColor BackgroundColor { get; set; }
-
-        internal string MatchEvaluator(Match m)
+        private string MatchEvaluator(Match m)
         {
             StringBuilder result = new StringBuilder();