--- a/src/NLog/Targets/ConsoleWordHighlightingRule.cs
+++ b/src/NLog/Targets/ConsoleWordHighlightingRule.cs
@@ -130,7 +130,7 @@ namespace NLog.Targets
                         }
                     }
 
-                    RegexOptions regexOptions = RegexOptions.None;
+                    RegexOptions regexOptions = RegexOptions.Compiled;
                     if (this.IgnoreCase)
                     {
                         regexOptions |= RegexOptions.IgnoreCase;