Fixes #980 

I've identified the expensive operations to be the `ForegroundColor` and `BackgroundColor` properties (both get and set), and the `TextWriter.Write()` and `TextWriter.WriteLine()` methods.

I've applied two performance enhancements: 
1) When there are no word highlighting rules, then we `WriteLine(message)` the message instead of doing a `Write(message)` followed by a `WriteLine()`. This saves one expensive `Write()` call per log.
2) We set the colors only if they must change (i.e. the matching rule's color is different).

To test the performance I built a program that logs 2000 messages and every 13th message is done with a different log level causing that message to appear in a different color. All other messages appear in the default console colors.

Before applying any of the improvements, the logging took 0.8 seconds on average. The `ConsoleTarget` takes about 0.25 seconds to log the same messages.

Change (1) yields an improvement of about 38 %, and change (2) yields an improvement of about 46 %.  
Together, the two changes decrease the running time from 0.8 seconds to about 0.33 seconds; a total improvement of about 59 %. It is also quite close to the `ConsoleTarget` time of 0.25 seconds.