@bhaeussermann 2016-01-01
I reverted the breaking change in `ConsoleWordHighlightingRule`. Thanks!