@304NotModified 2016-01-04
Hi, Thanks for this PR!

Just reviewed the code. two notes. Also, I think I think it would be nice to remove the regex fully when it's just a string replace, although I can be hard to keep that backwards-compatible. What do you think?