@bhaeussermann 2016-01-05
> Just reviewed the code. two notes. Also, I think I think it would be nice to remove the regex fully when it's just a string replace, although I can be hard to keep that backwards-compatible. What do you think?

I tested this and it yields no significant difference in performance, so it's probably not worth adding the additional complexity to the code.