@bhaeussermann 2016-01-05
Xamarin license issue again :(