@304NotModified 2016-01-05
My local machine seems to incorrectly unload the licenses.  :(