@304NotModified 2016-01-05
fixed. The deadlocked machines took some Xamarin licenses as my PC (deactivated now)