@bhaeussermann 2016-01-08
Thanks for the good proposals! I agree and I applied them.